requirements.txt
Pillow==2.1.0
requests==1.2.3
beautifulsoup4==4.2.1
